backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
redis>=5.2.0
httpx>=0.28.0
anthropic>=0.42.0
python-multipart>=0.0.18
pillow>=11.1.0
pytesseract>=0.3.13
pyzbar>=0.1.9
opencv-python-headless>=4.11.0
python-whois>=0.9.4
structlog>=24.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
boto3>=1.36.0
elasticsearch[async]>=8.17.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
httpx>=0.28.0
